Remove old websocket drafts
__Motivation__

Retain only the final draft for websocket and remove all previous versions.

__Modification__

Remove older websocket drafts.

__Result__

No more obsolete drafts.
