+++ codec-http/src/main/resources/META-INF/native-image/io.netty/codec-http/native-image.properties
@@ -13,4 +13,4 @@
 # under the License.
 
 Args = --initialize-at-build-time=io.netty \
-       --initialize-at-run-time=io.netty.handler.codec.http.HttpObjectEncoder,io.netty.handler.codec.http.websocketx.WebSocket00FrameEncoder,io.netty.handler.codec.http.websocketx.extensions.compression.DeflateDecoder,io.netty.handler.codec.compression.BrotliDecompressor
+       --initialize-at-run-time=io.netty.handler.codec.http.HttpObjectEncoder,io.netty.handler.codec.http.websocketx.extensions.compression.DeflateDecoder,io.netty.handler.codec.compression.BrotliDecompressor
