+++ codec-http/src/test/java/io/netty/handler/codec/http/websocketx/WebSocketServerProtocolHandlerTest.java
@@ -101,7 +101,7 @@ public class WebSocketServerProtocolHandlerTest {
                 .method(HttpMethod.GET)
                 .uri("/test")
                 .connection("Upgrade")
-                .version00()
+                .version13()
                 .upgrade("BogusSocket")
                 .build();
